% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTotalVolumeAndMaxQ.R
\name{addTotalVolumeAndMaxQ}
\alias{addTotalVolumeAndMaxQ}
\title{Add Total Volume and Max Q}
\usage{
addTotalVolumeAndMaxQ(qValues, events, eventnr, digitsV = 3,
  digitsMaxQ = 3)
}
\arguments{
\item{qValues}{vector of discharge values given in L/s}

\item{events}{event information as retrieved by \code{kwb.event::hsEvents}}

\item{eventnr}{integer vector of same length as \emph{qValues} giving the
number of the event to which the Q value belongs, as returned by
\code{kwb.event::hsEventNumber}.}

\item{digitsV}{number of decimal places for V in m3}

\item{digitsMaxQ}{number of decimal places for max. Q in L/s}
}
\value{
\emph{events} with columns \emph{V_m3} and \emph{maxQ_L_s} added
}
\description{
Add Total Volume and Max Q
}
