% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtMain.R
\name{completeTimeColumns}
\alias{completeTimeColumns}
\title{Complete Time Columns}
\usage{
completeTimeColumns(x, wanted = c("BerlinDateTime", "UTCOffset",
  "DateTimeUTC"))
}
\arguments{
\item{x}{data frame with time columns}

\item{wanted}{Default: c("BerlinDateTime", "UTCOffset", "DateTimeUTC")}
}
\value{
(Hopefully) data frame with columns \emph{BerlinDateTimeNoDST},
  \emph{BerlinDateTime}, \emph{UTCOffset}, \emph{DateTimeUTC},
}
\description{
Complete Time Columns
}
