% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_Q_Table.R
\name{convertQUnits}
\alias{convertQUnits}
\title{Convert Q in L/h to L/s, m3/s and m3/h}
\usage{
convertQUnits(hq)
}
\arguments{
\item{hq}{data frame containing a column \emph{Q_L_h} containing flows in L/h}
}
\value{
data frame with columns \emph{H_m, Q_L_s, Q_m3_s, Q_L_h, Q_m3_h}
}
\description{
Convert Q in L/h to L/s, m3/s and m3/h
}
