% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHQSeriesFromCSV.R
\name{correctHandCalculateQ}
\alias{correctHandCalculateQ}
\title{Correct Level with Offset and Calculate Q}
\usage{
correctHandCalculateQ(hdat, hoffset, DN)
}
\arguments{
\item{hdat}{data frame containing at least a column \emph{Hraw_m}, as e.g.
retrieved by \code{readLogger_Ori_MLog}}

\item{hoffset}{offset in m to be subtracted from raw levels before the HQ
relationship is applied}

\item{DN}{DN in mm, must be one of 150, 300}
}
\value{
data frame with additional columns \emph{H_m, Q_L_s}
}
\description{
Correct Level with Offset and Calculate Q
}
