% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirUploadedFiles.R
\name{dirFtpPath}
\alias{dirFtpPath}
\title{List Files on FTP Server}
\usage{
dirFtpPath(url, userpwd, full.names = FALSE)
}
\arguments{
\item{url}{base url in which to look for files}

\item{userpwd}{user and password, separated by colon ":"}

\item{full.names}{logical (default: \code{FALSE}). Determines whether to
return relative paths or the full URLs including the base \code{url}}
}
\value{
vector of urls or relative paths
}
\description{
List Files on FTP Server
}
