% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirUploadedFiles.R
\name{dirUploadedFiles}
\alias{dirUploadedFiles}
\title{List Uploaded Files}
\usage{
dirUploadedFiles(full.names = FALSE)
}
\arguments{
\item{full.names}{if \code{TRUE} the full absolute URLs are returned, 
otherwise only the relative paths.}
}
\value{
list with elements \emph{PN}, \emph{H}, \emph{RD}, \emph{F},
  \emph{LPR}, \emph{Q}, \emph{BPR} containing URLs to sampler files, water
  level files, rain data files, photos, laboratory protocol files, discharge
  files and operation protocol files, respectively, that are available at the
  DSWT server at sysprovide.de
}
\description{
You need to set the system environment variable "DSWT_FTP_LOGIN" to "user:pwd"
where "user" is the username and "pwd" the password for the account that is
allowed to access the FTP server where the files are stored.
}
