% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtConstants.R
\name{DSWT_BWB_CODE_TO_SITE_CODE}
\alias{DSWT_BWB_CODE_TO_SITE_CODE}
\title{Map BWB Rain Gauge Names to DSWT SiteCodes}
\usage{
DSWT_BWB_CODE_TO_SITE_CODE()
}
\value{
list with BWB names as element names and DSWT SiteCodes as elements
}
\description{
Map BWB Rain Gauge Names to DSWT SiteCodes
}
