% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DSWT_DICTIONARY_FILE}
\alias{DSWT_DICTIONARY_FILE}
\title{Path to Default Path Dictionary File}
\usage{
DSWT_DICTIONARY_FILE()
}
\value{
path to path dictionary file stored in the (installed) package
}
\description{
Default "dictionary" file describing the folder structure to be used in DSWT
}
