% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtConstants.R
\name{DSWT_FILE_TYPES}
\alias{DSWT_FILE_TYPES}
\title{List of File Type Definitions}
\usage{
DSWT_FILE_TYPES()
}
\value{
list mapping keywords to file name patterns
}
\description{
Containing e.g. a file name pattern
}
