% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DSWT_H_OFFSETS}
\alias{DSWT_H_OFFSETS}
\title{Water Level Offsets to be Subtracted from Measurement}
\usage{
DSWT_H_OFFSETS()
}
\value{
list of named elements with names corresponding to the monitoring
  site and the value corresponding to the offest in metres.
}
\description{
Water level offsets H_offset to be subtracted from measured level in order to
get the water level above the plume: H = H_raw - H_offset. H is then used to
calculate Q by using the relationship given by \code{\link{H_Q_Table}}
}
