% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtConstants.R
\name{DSWT_RAIN_GAUGES}
\alias{DSWT_RAIN_GAUGES}
\title{Rain Gauges Used in DSWT Project}
\usage{
DSWT_RAIN_GAUGES()
}
\value{
data frame with columns \emph{FUB_STATION}, \emph{FUB_SHORT}, 
  \emph{BWB_SHORT}
}
\description{
Rain Gauges Used in DSWT Project
}
