% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DSWT_SITES}
\alias{DSWT_SITES}
\title{Names of Sites Monitored in DSWT Project}
\usage{
DSWT_SITES()
}
\value{
named vector of SiteID values in ODM database
}
\description{
Names of Sites Monitored in DSWT Project
}
