% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DSWT_TIMESERIES}
\alias{DSWT_TIMESERIES}
\title{Foreign Keys Identifying Time Series in the ODM Data Model}
\usage{
DSWT_TIMESERIES()
}
\value{
list of lists
}
\description{
Foreign Keys Identifying Time Series in the ODM Data Model
}
