% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtMain.R
\name{dswtdir}
\alias{dswtdir}
\title{Path to Subfolder "DSWT" in tempdir()}
\usage{
dswtdir()
}
\value{
path to subfolder "DSWT" in tempdir()
}
\description{
Path to subfolder "DSWT" in tempdir(). If the subfolder does not yet exist
  it is created
}
