% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtAutoSampler.R
\name{getActionsFromAutoSamplerFile}
\alias{getActionsFromAutoSamplerFile}
\title{Get Actions from Autosampler File}
\usage{
getActionsFromAutoSamplerFile(pnFile, fileNumber = 1,
  remove.errors = FALSE)
}
\arguments{
\item{pnFile}{full path to ORI Auto sampler log files
PN_<yyyymmdd>_<station>.csv}

\item{fileNumber}{file number (will be included in the plot title). Useful if
this function is called in a sequence for multiple files.}

\item{remove.errors}{if TRUE, actions containing "Fehler" are removed}
}
\description{
Get Actions from Autosampler File
}
