% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtAutoSampler.R
\name{getActionsFromAutoSamplerFiles}
\alias{getActionsFromAutoSamplerFiles}
\title{Get Actions from Autosampler Files}
\usage{
getActionsFromAutoSamplerFiles(pnFiles)
}
\arguments{
\item{pnFiles}{full path(s) to ORI Auto sampler log files
PN_<yyyymmdd>_<station>.csv}
}
\description{
Get Actions from Autosampler Files
}
