% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevelFilesForSite.R
\name{getDswtFilePaths}
\alias{getDswtFilePaths}
\title{Get DSWT File Paths}
\usage{
getDswtFilePaths(srcdir, filetype, recursive = FALSE)
}
\arguments{
\item{srcdir}{source directory}

\item{filetype}{one of the file types contained in DSWT_FILE_TYPES}

\item{recursive}{search in subdirectories?}
}
\value{
vector of file paths
}
\description{
Browse for files of given type (DSWT-specific)
}
