% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtMain.R
\name{getHQSeriesFromCSV}
\alias{getHQSeriesFromCSV}
\title{Get H and Q Series from CSV File}
\usage{
getHQSeriesFromCSV(srcfile, DN, sep = "\\t", timeFormat = NULL,
  hoffset = 0.02, addTimeColumns = TRUE,
  additionalColumns = c("I_mA", "Battery_V"), ...)
}
\arguments{
\item{srcfile}{full path to csv file generated by radar probe}

\item{DN}{DN in mm. Must be one of 150, 300.}

\item{sep}{column separator. Default: Tabulator "\\t"}

\item{timeFormat}{format of timestamp. Default: "\%d.\%m.\%Y \%H:\%M"}

\item{hoffset}{level offset to be subtracted from the measured level in order
to get the level above the flume}

\item{addTimeColumns}{if TRUE, time columns containing local date and time
are added}

\item{additionalColumns}{names of additional columns to be imported. One of
c("I_mA", "Battery_V", "DeviceID")}

\item{\dots}{further arguments passed to readLogger_Ori_MLog, e.g.
\emph{stopOnMissingColumns}}
}
\value{
data frame with columns \emph{BerlinDateTimeNoDST} (no daylight
  saving time adjustment!), \emph{Hraw_m} (measured height in m), \emph{H_m }
  (corrected height [= measured height minus offset] in m),\emph{Q_L_s}
  (calculated discharge in L/s). If \emph{addTimeColumns} is TRUE the columns
  \emph{BerlinDateTime} and \emph{UTCOffset} will be added.
}
\description{
Get H and Q Series from CSV File
}
