% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtMain.R
\name{getLevelFilesForSite}
\alias{getLevelFilesForSite}
\title{Level Files for Site}
\usage{
getLevelFilesForSite(config, station)
}
\arguments{
\item{config}{configuration object (list) with elements "dictionaryFile" and
elements required by \code{\link[kwb.monitoring]{pathDictionary}}}

\item{station}{name of monitoring station}
}
\value{
vector of file paths
}
\description{
Level Files for Site
}
