% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{getLevelFilesInfo2}
\alias{getLevelFilesInfo2}
\title{Get Level Files Info 2}
\usage{
getLevelFilesInfo2(levelData)
}
\arguments{
\item{levelData}{data frame with columns \emph{myDateTime} (character),
\emph{file}, \emph{row}, as returned by \code{\link{readAllLevelFiles}}}
}
\value{
data frame with columns \code{file},\code{rows}, \code{min}, 
  \code{first}, \code{last}, \code{max}
}
\description{
Get Level Files Info 2
}
