% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_Q_Table.R
\name{H_Q_Table}
\alias{H_Q_Table}
\title{H-Q-Relationship Given by Manufacturer}
\usage{
H_Q_Table(DN)
}
\arguments{
\item{DN}{DN in mm, must be one of 150, 300}
}
\value{
data frame with columns \emph{H_m, Q_L_s, Q_m3_s, Q_L_h, Q_m3_h}
}
\description{
H-Q-Relationship Given by Manufacturer
}
