% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHQSeriesFromCSV.R
\name{H_to_Q}
\alias{H_to_Q}
\title{Calculate Q from Height Above flume}
\usage{
H_to_Q(H, DN)
}
\arguments{
\item{H}{height above flume in m}

\item{DN}{DN in mm, must be one of 150, 300}
}
\description{
Calculates Q from height h above flume as Q = a*H^b with a and be retrieved
  from linear regression between log(H) and log(Q) with H and Q values taken
  from manufacturer's table
}
