% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtMain.R
\name{insertUtcDateTimeColumn}
\alias{insertUtcDateTimeColumn}
\title{Insert DateTimeUTC Column}
\usage{
insertUtcDateTimeColumn(mydata)
}
\arguments{
\item{mydata}{data frame with column \code{BerlinDateTimeNoDST}}
}
\value{
mydata with additional column \emph{DateTimeUTC}
}
\description{
Insert DateTimeUTC Column
}
