% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{keyFields_DSWT}
\alias{keyFields_DSWT}
\title{Key Field Values in DSWT Project}
\usage{
keyFields_DSWT()
}
\value{
A list with each entry representing a time series, i.e. the 
  measurement of one variable at one site.
}
\description{
Key Field Values in DSWT Project
}
