% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtAutoSampler.R
\name{logEventToLabelInfo}
\alias{logEventToLabelInfo}
\title{Create Label and Colour Information from Event}
\usage{
logEventToLabelInfo(ereignis, x, colour.error = "red",
  colour.sample = "blue")
}
\value{
list with elements \emph{labels} and \emph{colours}
}
\description{
Create Label and Colour Information from Event
}
\keyword{internal}
