% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-sampler.R
\name{plotAutoSamplerActions}
\alias{plotAutoSamplerActions}
\title{Plot Autosampler Actions}
\usage{
plotAutoSamplerActions(
  actions,
  fileNumber,
  subevents.per.page = 3,
  density = 20,
  context = c(0, 0.12),
  evtSepTime = 30 * 60,
  ...
)
}
\description{
Plot Autosampler Actions
}
\keyword{internal}
