% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{prepareSingleVariableDataValuesForOdm}
\alias{prepareSingleVariableDataValuesForOdm}
\title{Prepare Single Variable Data Values for ODM}
\usage{
prepareSingleVariableDataValuesForOdm(
  dataFrame,
  colName,
  noDataValue = -9999,
  dbg = TRUE
)
}
\arguments{
\item{dataFrame}{data frame containing the data}

\item{colName}{column name}

\item{noDataValue}{value indicating "no data", default: -9999}

\item{dbg}{logical. If \code{TRUE}, debug messages are shown}
}
\description{
Prepare Single Variable Data Values for ODM
}
