% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHQSeriesFromCSV.R
\name{Q_to_H}
\alias{Q_to_H}
\title{Back-Calculate Height Above Flume From Discharge}
\usage{
Q_to_H(Q, DN)
}
\arguments{
\item{Q}{discharge Q in height above flume (in which unit?)}

\item{DN}{DN in mm, must be one of 150, 300}
}
\description{
back-calculates height H above flume from discharge Q.
  Q = a * H^b <=> H = (Q/a)^(1/b) with a and be retrieved from linear
  regression between log(H) and log(Q) with H and Q values taken from
  manufacturer's table
}
