% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtMain.R
\name{readAllLevelFiles}
\alias{readAllLevelFiles}
\title{Read all Level Files}
\usage{
readAllLevelFiles(levelFiles, dbg = TRUE)
}
\arguments{
\item{levelFiles}{vector of file paths}

\item{dbg}{if \code{TRUE}, debug messages are shown.}
}
\description{
Read all Level Files
}
