% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtAutoSampler.R
\name{readAndPlotAutoSamplerFiles}
\alias{readAndPlotAutoSamplerFiles}
\title{Read and Plot Autosampler Files}
\usage{
readAndPlotAutoSamplerFiles(filePaths,
  removePattern = "Power|Bluetooth|Modem|SMS|Sonde", to.pdf = TRUE,
  evtSepTime = 30 * 60)
}
\arguments{
\item{filePaths}{full path(s) to ORI Auto sampler log files
PN_<yyyymmdd>_<station>.csv}

\item{removePattern}{regular expression pattern matching logged actions to be
removed before plotting. Set to "" in order not to remove any action}

\item{to.pdf}{if TRUE, graphical output is directed to PDF}

\item{evtSepTime}{event separation time in seconds. Minimum time of 
"no signal" between two distinct events.}
}
\description{
Read and Plot Autosampler Files
}
