% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-sampler.R
\name{readDswtSamplerFileByName}
\alias{readDswtSamplerFileByName}
\title{Read File from Autosampler Used in Project DSWT}
\usage{
readDswtSamplerFileByName(samplerFile, bottlesToConsider)
}
\description{
Read File from Autosampler Used in Project DSWT
}
\keyword{internal}
