% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatEvents.R
\name{reformatEvents}
\alias{reformatEvents}
\title{Reformat Event List}
\usage{
reformatEvents(events)
}
\arguments{
\item{events}{event list as retrieved by \code{kwb.event::hsEvents}}
}
\value{
\emph{events} with \emph{tBeg} renamed \emph{Ereignisbeginn_UTC},
  \emph{tEnd} renamed \emph{Ereignisende_UTC}, \emph{dur} renamed
  \emph{Dauer_min}, \emph{pBefore} renamed \emph{Pause_davor_min} and
  \emph{pAfter} renamed \emph{Pause_danach_min} and original columns
  \emph{iBeg} and \emph{iEnd} removed
}
\description{
Reformat event list: convert to minutes and rename columns
}
