% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtValidate.R
\name{toLevelFileInfo}
\alias{toLevelFileInfo}
\title{Create Information on Water Level File}
\usage{
toLevelFileInfo(filePath, timestamps)
}
\arguments{
\item{filePath}{path to water level file}

\item{timestamps}{vector of timestamps (read from the file?)}
}
\description{
Create Information on Water Level File
}
