% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_h-q.R
\name{validate_HQ_relationship}
\alias{validate_HQ_relationship}
\title{validate_HQ_relationship}
\usage{
validate_HQ_relationship(DN)
}
\description{
Validate HQ relationship by plotting Q versus H from manufacturer table
  as points and the regression line given by the derived formula
}
\keyword{internal}
