% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtMain.R
\name{writeEventListToCSV}
\alias{writeEventListToCSV}
\title{Write Event List to CSV File}
\usage{
writeEventListToCSV(events, csv, sep = ";", dec = ",")
}
\arguments{
\item{events}{data frame containing event data}

\item{csv}{full path to csv file}

\item{sep}{column separator. Default: ";"}

\item{dec}{decimal character. Default: "."}
}
\description{
Write Event List to CSV File
}
