% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dswtMain.R
\name{writeHQSeriesToCSV}
\alias{writeHQSeriesToCSV}
\title{Write H-Q-Series to CSV File}
\usage{
writeHQSeriesToCSV(hqSeries, csv, sep = ";", dec = ",")
}
\arguments{
\item{hqSeries}{data frame containing HQ time series}

\item{csv}{full path to csv file}

\item{sep}{column separator. Default: ";"}

\item{dec}{decimal character. Default: "."}
}
\description{
Write H-Q-Series to CSV File
}
