% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_date-time-columns.R
\name{insertLocalDateTimeColumns}
\alias{insertLocalDateTimeColumns}
\title{Insert LocalDateTime Columns}
\usage{
insertLocalDateTimeColumns(mydata)
}
\arguments{
\item{mydata}{data frame with character column \emph{BerlinDateTimeNoDST}}
}
\value{
data frame with additional columns \emph{BerlinDateTime} (character),
  \emph{UTCOffset} (numeric)
}
\description{
Insert LocalDateTime Columns
}
